requirements.txt
# Package dependencies
cloud-sql-python-connector[pg8000]
google-cloud-alloydb-connector[pg8000]
cloud-sql-python-connector[pytds]
sqlalchemy-pytds
google-genai
google-cloud-secret-manager
google-cloud-bigtable
google-cloud-spanner
sqlalchemy-spanner
pandas
pandas-gbq
pymysql
sqlalchemy
pyaml_env
pyarrow
protobuf
absl-py
tabulate
grpcio
grpcio-tools
aiologger
jsonschema
ratelimit
backoff
GitPython
sqlparse
redis
anthropic[vertex]
sqlglot
pymongo
mongomock
rich
google-adk
